+++ freeRtos_uart/timer.h
@@ -8,4 +8,4 @@
 void timer0_init(void);
 uint32_t millis(void);
 uint32_t micros(void);
-void delay(float ms);
+void delay(double ms);
